[tests] Move type map rewrite tests to net11
Run the JNI rewriter cases that directly consume the trimmable type map generator in the existing net11 integration test assembly. This removes the accidental transitive reference from the net10 Xamarin.Android.Build.Tests project.
