+++ tests/Microsoft.Android.Sdk.TrimmableTypeMap.IntegrationTests/MockBuildEngine.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections;
+using System.Collections.Generic;
 using Microsoft.Build.Framework;
 
 namespace Microsoft.Android.Sdk.TrimmableTypeMap.IntegrationTests;
@@ -9,6 +10,13 @@ namespace Microsoft.Android.Sdk.TrimmableTypeMap.IntegrationTests;
 /// </summary>
 sealed class MockBuildEngine : IBuildEngine
 {
+	readonly IList<BuildWarningEventArgs>? warnings;
+
+	public MockBuildEngine (IList<BuildWarningEventArgs>? warnings = null)
+	{
+		this.warnings = warnings;
+	}
+
 	public bool ContinueOnError => false;
 	public int LineNumberOfTaskNode => 0;
 	public int ColumnNumberOfTaskNode => 0;
@@ -18,5 +26,5 @@ sealed class MockBuildEngine : IBuildEngine
 	public void LogCustomEvent (CustomBuildEventArgs e) { }
 	public void LogErrorEvent (BuildErrorEventArgs e) { }
 	public void LogMessageEvent (BuildMessageEventArgs e) { }
-	public void LogWarningEvent (BuildWarningEventArgs e) { }
+	public void LogWarningEvent (BuildWarningEventArgs e) => warnings?.Add (e);
 }
